requirements.txt
Flask===2.3.3
SQLAlchemy===2.0.20
flask_sqlalchemy===3.1.1
Flask_Migrate===4.0.5
PyJWT===1.7.1
pytest===7.4.2
pytest_xdist===3.3.1
pytest_cov===4.1.0
coverage===7.3.1
Flask_Cors===4.0.0
